requirements_get.txt
asgiref==3.2.3
blis==0.4.1
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
cymem==2.0.3
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.2
django-import-export==2.0.1
django-summernote==0.8.11.6
djangorestframework==3.11.0
en-core-web-sm==2.2.5
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
jdcal==1.4.1
MarkupPy==1.14
murmurhash==1.0.2
numpy==1.18.1
odfpy==1.4.1
openpyxl==3.0.3
plac==1.1.3
preshed==3.0.2
psycopg2-binary==2.8.4
PyMySQL==0.9.3
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
six==1.14.0
spacy==2.2.3
sqlparse==0.3.0
srsly==1.0.1
stripe==2.42.0
tablib==0.14.0
textstat==0.6.0
thinc==7.3.1
tqdm==4.42.0
urllib3==1.25.8
wasabi==0.6.0
xlrd==1.2.0
XlsxWriter==1.2.7
xlwt==1.3.0
zipp==2.1.0
